VRMExporterVaildator.cs に分割
