+++ Assets/VRM/UniVRM/Editor/Format/VRMExporterVaildator.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bc233cbadb897eb4886de9927bee9fc2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
